etl/feeds/requirements.txt
pandas
openpyxl
requests
lxml
html5lib
PyYAML
beautifulsoup4
playwright
psycopg2-binary
feedparser
libretranslatepy
